Move the Repair Task classes to new package
Changes:
- Create new subpackage task under malfunction
- Move the Repair Task classes to the new location
- Update references

Ref: #1298
