+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/task/util/MetaTaskUtil.java
@@ -25,6 +25,7 @@ import com.mars_sim.core.building.function.task.OptimizeSystemMeta;
 import com.mars_sim.core.building.function.task.ProduceFoodMeta;
 import com.mars_sim.core.building.function.task.ToggleFuelPowerSourceMeta;
 import com.mars_sim.core.building.task.MaintainBuildingMeta;
+import com.mars_sim.core.malfunction.task.RepairMalfunctionMeta;
 import com.mars_sim.core.manufacture.task.ManufacturingMetaTask;
 import com.mars_sim.core.person.ai.task.meta.AnalyzeMapDataMeta;
 import com.mars_sim.core.person.ai.task.meta.ConnectOnlineMeta;
@@ -41,7 +42,6 @@ import com.mars_sim.core.person.ai.task.meta.PlayHoloGameMeta;
 import com.mars_sim.core.person.ai.task.meta.ReadMeta;
 import com.mars_sim.core.person.ai.task.meta.RecordActivityMeta;
 import com.mars_sim.core.person.ai.task.meta.RelaxMeta;
-import com.mars_sim.core.person.ai.task.meta.RepairMalfunctionMeta;
 import com.mars_sim.core.person.ai.task.meta.ReportMissionControlMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewJobReassignmentMeta;
 import com.mars_sim.core.person.ai.task.meta.ReviewMissionPlanMeta;
